+++ doc/doc_app.py
@@ -13,11 +13,10 @@ WSGIRequestHandler.log_request = lambda *args, **kwargs: None
 
 
 def create_doc_app(sphinx_app):
-    conf_path = (
-        pathlib.Path(__file__).parent.parent / "dev" / "conf" / "canaille-memory.toml"
-    )
+    conf_path = pathlib.Path(__file__).parent.parent / "dev" / "conf" / "canaille.toml"
     with open(conf_path) as fd:
         conf = dict(tomlkit.load(fd))
+    conf["CANAILLE"]["DATABASE"] = "memory"
     conf["CANAILLE"]["SECRET_KEY"] = "doc"
     conf["CANAILLE"]["SMTP"] = {"HOST": "localhost"}
     conf["CANAILLE"]["LANGUAGE"] = sphinx_app.config["language"]